requirements.txt
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0
numpy>=1.26
polars>=1.5.0
pyarrow>=16.0.0
scikit-learn>=1.5.0
pyyaml>=6.0.1
rich>=13.7.0
tensorboard>=2.17.0
